refactor(ffi): extract diff and pack entrypoints
